+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Corrected normalization behavior for scene thumbnails and annotation projects [#5569](https://github.com/raster-foundry/raster-foundry/pull/5569)
+- Restored coordinate space to 4096x4096 in label MVT [#5569](https://github.com/raster-foundry/raster-foundry/pull/5569)
 
 ## [1.62.0] - 2021-04-14
 ### Changed