+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed campaign STAC exports so that classes are added to label STAC item and label data [#5570](https://github.com/raster-foundry/raster-foundry/pull/5570)
 
 ## [1.62.0] - 2021-04-14
 ### Changed